refactor: Move onAuthStateChanged to FirebaseApp component
